[SYSTEMDS-2907] Fix memory estimates dense and sparse matrices
The dense and sparse memory estimates are used for multiple purposes
including selecting distributed operations if needed. A recent change of
these memory estimates introduced a few issues which lead overflows
estimates memory of ultra-sparse matrices (which dense estimate exceed
int max, and in some cases long max).

* Fix int overflows on estimating memory of dense and spare matrices
* Fix incorrect estimates used by the compiler (header in wrong place)
* Fix missing long overflow checks

+++ src/main/java/org/apache/sysds/runtime/compress/colgroup/Dictionary.java
@@ -63,7 +63,7 @@ public class Dictionary extends ADictionary {
 
 	protected static long getInMemorySize(int valuesCount) {
 		// object + values array
-		return 16 + MemoryEstimates.doubleArrayCost(valuesCount);
+		return 16 + (long)MemoryEstimates.doubleArrayCost(valuesCount);
 	}
 
 	@Override
